Refaktorer og begrens visning av spm på ikkeSammeForelder spm
