+++ src/søknad/steg/4-barnasbosted/ikkesammeforelder/BoddSammenFør.tsx
@@ -1,0 +1,76 @@
+import { IForelder } from '../../../../models/forelder';
+import React, { FC } from 'react';
+import { useIntl } from 'react-intl';
+import KomponentGruppe from '../../../../components/gruppe/KomponentGruppe';
+import JaNeiSpørsmål from '../../../../components/spørsmål/JaNeiSpørsmål';
+import { boddSammenFør } from '../ForeldreConfig';
+import Datovelger, {
+  DatoBegrensning,
+} from '../../../../components/dato/Datovelger';
+import { ESvar, ISpørsmål, ISvar } from '../../../../models/spørsmålogsvar';
+import { hentTekst } from '../../../../utils/søknad';
+import { hentBooleanFraValgtSvar } from '../../../../utils/spørsmålogsvar';
+
+interface Props {
+  forelder: IForelder;
+  settForelder: Function;
+}
+const BoddSammenFør: FC<Props> = ({ forelder, settForelder }) => {
+  const intl = useIntl();
+
+  const settHarBoddsammenFør = (spørsmål: ISpørsmål, valgtSvar: ISvar) => {
+    const nyForelder = {
+      ...forelder,
+      [boddSammenFør.søknadid]: {
+        spørsmålid: spørsmål.søknadid,
+        svarid: valgtSvar.id,
+        label: hentTekst(spørsmål.tekstid, intl),
+        verdi: hentBooleanFraValgtSvar(valgtSvar),
+      },
+    };
+
+    if (valgtSvar.id === ESvar.NEI) {
+      delete nyForelder.flyttetFra;
+    }
+
+    settForelder(nyForelder);
+  };
+
+  return (
+    <>
+      <KomponentGruppe>
+        <JaNeiSpørsmål
+          spørsmål={boddSammenFør}
+          onChange={(spørsmål, svar) => settHarBoddsammenFør(spørsmål, svar)}
+          valgtSvar={forelder.boddSammenFør?.verdi}
+        />
+      </KomponentGruppe>
+      {forelder.boddSammenFør?.verdi && (
+        <KomponentGruppe>
+          <Datovelger
+            settDato={(e: Date | null) => {
+              e !== null &&
+                settForelder({
+                  ...forelder,
+                  flyttetFra: {
+                    label: intl.formatMessage({
+                      id: 'barnasbosted.normaltekst.nårflyttetfra',
+                    }),
+                    verdi: e,
+                  },
+                });
+            }}
+            valgtDato={
+              forelder.flyttetFra && forelder.flyttetFra.verdi
+                ? forelder.flyttetFra.verdi
+                : undefined
+            }
+            tekstid={'barnasbosted.normaltekst.nårflyttetfra'}
+            datobegrensning={DatoBegrensning.TidligereDatoer}
+          />
+        </KomponentGruppe>
+      )}
+    </>
+  );
+};
+export default BoddSammenFør;
